ADD: add the class `CbetaMissingCharacter`
Motivation:
The CbetaMissingCharacter is used as an database of all the character that is missing or in different form shown in the cbeta file. It database is in the form of Python dictionary and it can be updated by extracting information from cbeta file.

Modification:
1. Add package `cbeta_missing`
2. Add module `cbeta_missing_character`
3. Add class `CbetaMissingCharacter`
4. realize part of the planed methods of the class

Result:
Realize part of the methods, there are some methods to do and a unittest for the class